requirements.txt
Django==4.2.10
djangorestframework==3.14.0
drf-spectacular==0.27.1
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
python-decouple==3.8
django-filter==23.5
django-cors-headers==4.3.1
gunicorn==21.2.0
uvicorn==0.25.0
asgiref==3.7.1
djangorestframework-simplejwt==5.3.1
cryptography==41.0.7
channels==4.0.0
channels-redis==4.1.0
locust==2.24.1
psutil==5.9.8
bandit==1.7.7
safety==3.0.1